requirements.txt
markdown
pyyaml
bs4
stravalib
jinja2
watchdog
py-markdown-table